Aplicación funcionando correctamente.

+++ app/views/VistaListar.php
@@ -27,8 +27,10 @@
         </td>
         <td>
           <a href="?c=Tareas&a=Estado&id=<?= $task['id']?>"><button class="btn btn-primary right botonwidth"><?php echo $task['estado'];?></button></a>
-          <a href="?c=Tareas&a=Edit&id=<?= $task['id']?>"><button class="btn btn-primary right botonwidth">Modificar</button></a>
-          <a href="?c=Tareas&a=Delete&id=<?= $task['id']?>"><button class="btn btn-danger right botonwidth">Eliminar</button></a>         
+            <?php if (isset($_SESSION['tipousuario']) && $_SESSION['tipousuario'] == 'A'):?>           
+                    <a href="?c=Tareas&a=Edit&id=<?= $task['id']?>"><button class="btn btn-primary right botonwidth">Modificar</button></a>
+                    <a href="?c=Tareas&a=Delete&id=<?= $task['id']?>"><button class="btn btn-danger right botonwidth">Eliminar</button></a>         
+             <?php endif; ?>       
           <?php endforeach; ?>
            
         </td>
